added user verification for delete and update functions

+++ books/views.py
@@ -38,6 +38,27 @@ def list_books(request: HttpRequest):
 
     return render(request, 'books/home.html', context={"books": books})
 
+
+def list_user_books(request: HttpRequest, user_pk: int):
+
+    sort = request.GET.get("sort")
+    books = Book.objects.filter(user_id=user_pk).all().order_by("pk")
+
+
+    if sort == "title=asc":
+        books = Book.objects.filter(user_id=user_pk).all().order_by("title")
+    if sort == "title=desc":
+        books = Book.objects.filter(user_id=user_pk).all().order_by("-title")
+    if sort == "default":
+        books = Book.objects.filter(user_id=user_pk).all().order_by("pk")
+    if sort == "author=desc":
+        books = Book.objects.filter(user_id=user_pk).all().order_by("-author")
+    if sort == "author=asc":
+        books = Book.objects.filter(user_id=user_pk).all().order_by("author")
+
+    return render(request, 'books/home.html', context={"books": books})
+
+
 @login_required
 def create_book(request: HttpRequest):
     if request.method == 'POST':
@@ -56,29 +77,35 @@ def create_book(request: HttpRequest):
         list1 = [10, 20, 30, 40, 50]
         return render(request, 'books/book_form.html', context={'form': form, 'list1': list1})
 
-
+@login_required
 def delete_book(request: HttpRequest, pk: int):
     book = get_object_or_404(Book, pk=pk)
-    if request.method == 'POST':
-        book.delete()
-        return redirect('home')
+
+    if request.user.pk == book.user.pk:
+        if request.method == 'POST':
+            book.delete()
+            return redirect('home')
+        else:
+            return render(request, 'books/book_confirm_delete.html', context={'book': book})
     else:
-        return render(request, 'books/book_confirm_delete.html', context={'book': book})
+        return HttpResponse("Fuck you")
 
-
+@login_required
 def update_book(request: HttpRequest, pk: int):
     book = get_object_or_404(Book, pk=pk)
+    if request.user.pk == book.user.pk:
     # book = Book.objects.get(pk=pk)
-    if request.method == 'POST':
-        # detaliile book-ului care au fost trimise de form folosind HTTP POST request, se afla in request.POST ca un dictionar.
-        book_instance = BookForm(request.POST, instance=book)
-        if book_instance.is_valid():
-            # aici se updateaza un book in baza de date!
-            book_instance.save()
-            return redirect('home')
-        # return HttpResponse('ÜBERRASCHUNG!')
-    else:
-        # in cazul asta, request-ul poate fi GET, PUT, PATCH, DELETE, etc
-        form = BookForm(instance=book)
-        list1 = [10, 20, 30, 40, 50]
-        return render(request, 'books/update_book_form.html', context={'form': form})
+        if request.method == 'POST':
+            # detaliile book-ului care au fost trimise de form folosind HTTP POST request, se afla in request.POST ca un dictionar.
+            book_instance = BookForm(request.POST, instance=book)
+            if book_instance.is_valid():
+                # aici se updateaza un book in baza de date!
+                book_instance.save()
+                return redirect('home')
+            # return HttpResponse('ÜBERRASCHUNG!')
+        else:
+            # in cazul asta, request-ul poate fi GET, PUT, PATCH, DELETE, etc
+            form = BookForm(instance=book)
+            list1 = [10, 20, 30, 40, 50]
+            return render(request, 'books/update_book_form.html', context={'form': form})
+    return HttpResponse("Fuck you")
